pyproject.toml
[project]
name = "statextract"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "httpx",
    "beautifulsoup4",
    "joblib",
    "anthropic",
    "pydantic",
    # "pyyaml",
    "instructor",
    "pymupdf",
    "pymupdf4llm",
    "perscache",
    "opentelemetry-instrumentation-anthropic>=0.32.2",
    "pyalex>=0.15.1",
    "fastapi[standard]>=0.115.2",
    "pypcurve>=0.1.1",
    "pillow>=11.0.0",
    "aiodbm>=0.4.2",
]
